Fix annoying cursor movement on grid click
This fix does two things, first it adds a supplementary text property to
the image that makes point go all the way to the end whenever it enters
the image, second the active region is deactivated to avoid it being
visible if a click managed to slip beyond the image and activated it.

The latter feels a bit overkill, but should be fine when kept to this
buffer only. Closes #1.