Implement buffer-local playing state
This change revealed that timers would have become "buffer-local" if you
were to not adjust the play/pause semantics.  I'm not sure on the
details behind it, but pausing apparently tried canceling the timer and
playing introduced a new one which resulted in one timer per buffer that
affected the other buffer as well.

Might have been easier to just use one timer per buffer.  Oh well.  That
would be a possible #3.  Closes #2.